services/reading_indexer/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sentence-transformers==3.3.0
psycopg2-binary==2.9.10
numpy==1.26.4
python-dotenv==1.0.1
